Reorganize into two libraries with different build options. Issue: only one library can be linked by `pc` at once... but we want different ones for the two libs
