+++ native/hb_beamr/lib/stub.c
@@ -1,0 +1,1 @@
+#include "wasm_export.h"
