+++ native/hb_beamrc/lib/include/stub.h
