+++ native/hb_beamrc/lib/test/helpers.h
