+++ native/hb_beamrc/lib/test/test_compile.c
