+++ rebar.config
@@ -39,26 +39,40 @@
 {overrides, []}.
 
 {pre_hooks, [
-	{compile, "make -C \"${REBAR_ROOT_DIR}/native/common/wasm\" CMAKE_BUILD_DIR=\"${REBAR_ROOT_DIR}/_build/hb_wasm\""}
+    {compile, "mkdir -p port_libs"},
+    {compile, "make -C \"${REBAR_ROOT_DIR}/native/hb_beamr/lib\"  CMAKE_BUILD_DIR=\"${REBAR_ROOT_DIR}/_build/port_libs/hb_beamr\""},
+    {compile, "make -C \"${REBAR_ROOT_DIR}/native/hb_beamrc/lib\" CMAKE_BUILD_DIR=\"${REBAR_ROOT_DIR}/_build/port_libs/hb_beamrc\""}
 ]}.
 
 {port_env, [
 	{"(linux|darwin|solaris)", "CFLAGS", "$CFLAGS" 
-        " -I${REBAR_ROOT_DIR}/_build/hb_wasm/wamr-src/core/iwasm/include"
-        " -I${REBAR_ROOT_DIR}/_build/hb_wasm/wamr-src/core/deps/llvm/build/include"
-        " -I${REBAR_ROOT_DIR}/native/common/wasm/include" },
+        " -I${REBAR_ROOT_DIR}/native/common/driver/include"
+        " -I${REBAR_ROOT_DIR}/_build/port_libs/hb_beamr/wamr-src/core/iwasm/include"
+        % The headers are identical, so no need to use the other directory
+        % " -I${REBAR_ROOT_DIR}/_build/port_libs/hb_beamrc/wamr-src/core/iwasm/include"
+    },
 	{"(linux|darwin|solaris)", "LDFLAGS", "$LDFLAGS"
-        " -L${REBAR_ROOT_DIR}/_build/hb_wasm"
-        " -Wl,-rpath,${REBAR_ROOT_DIR}/_build/hb_wasm"
-        " -L${REBAR_ROOT_DIR}/_build/hb_wasm/wamr-src/core/deps/llvm/build/lib" },
-	{"(linux|darwin|solaris)", "LDLIBS", "$LDLIBS -lhb_wasm -lvmlib -laotclib -lei -lpthread -lm"}
+        % Dynamic linking directory for hb_beamr_lib
+        " -Wl,-rpath,${REBAR_ROOT_DIR}/_build/port_libs/hb_beamr"
+        " -L${REBAR_ROOT_DIR}/_build/port_libs/hb_beamr"
+        % Dynamic linking directory for hb_beamrc_lib
+        " -Wl,-rpath,${REBAR_ROOT_DIR}/_build/port_libs/hb_beamrc"
+        " -L${REBAR_ROOT_DIR}/_build/port_libs/hb_beamrc"
+    },
+	{"(linux|darwin|solaris)", "LDLIBS", "$LDLIBS"
+        " -lvmlib -lei -lpthread -lm"
+        % Only one of these lines can be enabled, or it will cause conflicts!
+        % " -lhb_beamr_lib"
+        " -lhb_beamrc_lib -laotclib"
+    }
 ]}.
 
 {post_hooks, [
 	{"(linux|darwin|solaris)", clean, "rm -rf \"${REBAR_ROOT_DIR}/_build\" \"${REBAR_ROOT_DIR}/priv\""},
 	{"(linux|darwin|solaris)", compile, "echo 'Post-compile hooks executed'"},
-    { compile, "rm -f native/hb_beamr/driver/src/*.o native/hb_beamr/driver/src/*.d"},
-    { compile, "rm -f native/hb_beamr/compiler/src/*.o native/hb_beamr/compiler/src/*.d"},
+    { compile, "rm -f native/hb_beamr/src/*.o native/hb_beamr/src/*.d"},
+    { compile, "rm -f native/hb_beamrc/src/*.o native/hb_beamrc/src/*.d"},
+    { compile, "rm -f native/common/driver/src/*.o native/common/driver/src/*.d"},
     { compile, "mkdir -p priv/html"},
     { compile, "cp -R src/html/* priv/html"}
 ]}.
@@ -76,16 +90,16 @@
 
 {port_specs, [
 	{"priv/hb_beamr.so", [
-        "native/hb_beamr/driver/src/hb_beamr.c",
-        "native/hb_beamr/driver/src/hb_driver.c",
-        "native/hb_beamr/driver/src/hb_helpers.c",
-        "native/hb_beamr/driver/src/hb_logging.c",
-        "native/hb_beamr/driver/src/hb_wasm.c"
+        "native/hb_beamr/src/hb_beamr.c",
+        "native/hb_beamr/src/hb_helpers.c",
+        "native/hb_beamr/src/hb_wasm.c",
+        "native/common/driver/src/hb_driver.c",
+        "native/common/driver/src/hb_logging.c"
     ]},
     {"priv/hb_beamrc.so", [
-        "native/hb_beamr/compiler/src/hb_beamrc.c",
-        "native/hb_beamr/compiler/src/hb_driver.c",
-        "native/hb_beamr/compiler/src/hb_logging.c"
+        "native/hb_beamrc/src/hb_beamrc.c",
+        "native/common/driver/src/hb_driver.c",
+        "native/common/driver/src/hb_logging.c"
     ]}
 ]}.
